refactor: 실제 도메인에 없는 단순 타입용 클래스들의 패키지 변경
