+++ src/main/java/nextstep/courses/type/CapacityUnit.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.type;
 
 public enum CapacityUnit {
     KB("KB");
