+++ src/test/java/nextstep/courses/domain/SessionImageTest.java
@@ -3,7 +3,7 @@ package nextstep.courses.domain;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import static nextstep.courses.domain.ImageExtension.*;
+import static nextstep.courses.type.ImageExtension.*;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
